Simplify implementation of chunked file storage.
Does away with all the complication of separate reader and writer threads.
Will now read the input until one full-size chunk is ready to upload, then
block the input until the chunk has been stored, and repeat.  This was the
practical outcome anyway in all situations where the read thread is reliably
much faster than the upload thread, which was always, when the data was
reading from a local dist.